Integrate model migrations
